+++ changelog.md
@@ -1,9 +1,10 @@
-## Forestry 2.2.4
+## Forestry 2.3.0
 - Fix saplings with default genomes not generating leaves properly (#144)
 - Add API for IBeekeepingLogic to change throttle ticks
 - Fix NBT loading bug in BeekeepingLogic to load queen max health
 - Make parts of JEI plugin public
 - Fix IBeeModifier.modifyAging and IIndividualLiving.age
+- Fix bee territory being too small on East and West (#162)
 
 ## Forestry 2.2.3
 - Add Cathemeral bee activity type, which is a random activity period of 12000 ticks that varies based on position. Check times using /forestry bee cathemeral